+++ lib/graphql-pundit/field.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'graphql'
-require 'pry'
 require 'graphql-pundit/authorization'
 require 'graphql-pundit/scope'
 
@@ -11,8 +10,6 @@ module GraphQL
     class Field < GraphQL::Schema::Field
       prepend GraphQL::Pundit::Scope
       prepend GraphQL::Pundit::Authorization
-
-      current_user :me
     end
   end
 end
