+++ spec/graphql-pundit/scope_spec.rb
@@ -7,21 +7,48 @@ RSpec.describe GraphQL::Pundit::Scope do
 
   subject { ScopeTest.new([1, 2, 3, 22, 48]) }
 
-  context 'without authorization' do
-    context 'inferred scope' do
-      let(:field) do
-        Field.new(name: :to_a,
-                  type: String,
-                  before_scope: true,
-                  null: true)
+  context 'one-line field definition' do
+    let(:field) do
+      Field.new(name: :to_a,
+                type: String,
+                before_scope: before_scope,
+                authorize: authorize,
+                null: true)
+    end
+
+    context 'without authorization' do
+      let(:authorize) { nil }
+      context 'inferred scope' do
+        let(:before_scope) { true }
+        it 'filters the list' do
+          expect(result).to match_array([1, 2, 3])
+        end
+
+        context 'scope from model' do
+          subject { ScopeTestDataset.new([1, 2, 3, 22, 48]) }
+
+          it 'filters the list' do
+            expect(result).to match_array([1, 2, 3])
+          end
+        end
       end
 
-      it 'filters the list' do
-        expect(result).to match_array([1, 2, 3])
+      context 'explicit scope proc' do
+        let(:before_scope) do
+          lambda do |scope, _args, _ctx|
+            scope.where { |e| e > 20 }
+          end
+        end
+
+        it 'filters the list' do
+          expect(result).to match_array([22, 48])
+        end
       end
 
-      context 'scope from model' do
-        subject { ScopeTestDataset.new([1, 2, 3, 22, 48]) }
+      context 'explicit scope class' do
+        let(:before_scope) do
+          ScopeTestPolicy
+        end
 
         it 'filters the list' do
           expect(result).to match_array([1, 2, 3])
@@ -29,93 +56,47 @@ RSpec.describe GraphQL::Pundit::Scope do
       end
     end
 
-    context 'explicit scope proc' do
-      let(:field) do
-        Field.new(name: :to_a,
-                  type: String,
-                  before_scope: (lambda do |scope, _args, _ctx|
-                    scope.where { |e| e > 20 }
-                  end),
-                  null: true)
+    context 'with authorization' do
+      let(:authorize) { true }
+      context 'inferred scope' do
+        let(:before_scope) { true }
+
+        it 'returns nil' do
+          expect(result).to eq(nil)
+        end
+
+        context 'scope from model' do
+          subject { ScopeTestDataset.new([1, 2, 3, 22, 48]) }
+
+          it 'returns nil' do
+            expect(result).to eq(nil)
+          end
+        end
       end
 
-      it 'filters the list' do
-        expect(result).to match_array([22, 48])
-      end
-    end
-
-    context 'explicit scope class' do
-      let(:field) do
-        Field.new(name: :to_a,
-                  type: String,
-                  before_scope: ScopeTestPolicy,
-                  null: true)
-      end
-
-      it 'filters the list' do
-        expect(result).to match_array([1, 2, 3])
-      end
-    end
-  end
-
-  context 'with authorization' do
-    context 'inferred scope' do
-      let(:field) do
-        Field.new(name: :to_a,
-                  type: String,
-                  before_scope: true,
-                  authorize: true,
-                  null: true)
-      end
-
-      it 'returns nil' do
-        expect(result).to eq(nil)
-      end
-
-      context 'scope from model' do
-        subject { ScopeTestDataset.new([1, 2, 3, 22, 48]) }
-        let(:field) do
-          Field.new(name: :to_a,
-                    type: String,
-                    before_scope: true,
-                    authorize: true,
-                    policy: ScopeTestPolicy,
-                    null: true)
+      context 'explicit scope proc' do
+        let(:before_scope) do
+          lambda do |scope, _args, _ctx|
+            # :nocov:
+            # This is supposed to not be run
+            scope.where { |e| e > 20 }
+            # :nocov:
+          end
         end
 
         it 'returns nil' do
           expect(result).to eq(nil)
         end
       end
-    end
 
-    context 'explicit scope proc' do
-      let(:field) do
-        Field.new(name: :to_a,
-                  type: String,
-                  before_scope: (lambda do |scope, _args, _ctx|
-                    scope.where { |e| e > 20 }
-                  end),
-                  authorize: true,
-                  null: true)
-      end
+      context 'explicit scope class' do
+        let(:before_scope) do
+          ScopeTestPolicy
+        end
 
-      it 'returns nil' do
-        expect(result).to eq(nil)
-      end
-    end
-
-    context 'explicit scope class' do
-      let(:field) do
-        Field.new(name: :to_a,
-                  type: String,
-                  before_scope: ScopeTestPolicy,
-                  authorize: true,
-                  null: true)
-      end
-
-      it 'returns nil' do
-        expect(result).to eq(nil)
+        it 'returns nil' do
+          expect(result).to eq(nil)
+        end
       end
     end
   end
